requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
feedparser>=6.0.10
google-generativeai>=0.3.0
apscheduler>=3.10.4
httpx>=0.26.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
jinja2>=3.1.3
tenacity>=8.2.3
python-dateutil>=2.8.2
elevenlabs>=0.2.27
